Rework TK Army, add Night Fighting variant
